classify/postgres: §7.6 escalate_unknown_functions knob + corpus
Replaces the Task-13 stub in ast_dml.go with a real implementation in
escalation.go. Refactors the FuncCall AST walker out of ast_unsafe_io.go
into a shared walkFuncCalls helper so the unsafe-IO detector and the
escalation detector share a single traversal.

collectFuncCallsAny returns true if any FuncCall in the SelectStmt has
a canonical lowercase name not in the allowlist. Schema-qualified names
are matched against their dotted lowercase form (e.g. "public.now");
bare calls match the lowercase identifier. classifySelect appends a
group=procedural (no Subtype) effect when EscalateUnknownFunctions is
true and the walker reports an unknown call.

Adds escalation_test.go (allowlist-hit / -miss, schema-qualified,
nested args, mixed, WHERE-clause, case-insensitive, end-to-end via
classifySelect) and corpus rows 0150 (allowlisted now() stays read)
and 0151 (unknown my_volatile() escalates to procedural primary).

Cross-compiles clean for windows + darwin; full go test ./... passes.
