+++ internal/db/classify/postgres/corpus/0151-select-unknown-fn-escalates.yaml
@@ -1,0 +1,18 @@
+name: select-unknown-fn-escalates
+spec_ref: §7.6 escalate_unknown_functions knob
+description: |
+  SELECT my_volatile() with EscalateUnknownFunctions=true; my_volatile is
+  not in the allowlist (now, current_timestamp). The classifier appends a
+  procedural effect alongside the read; canonical ordering promotes
+  procedural (High severity) ahead of read (Low) so the primary becomes
+  procedural.
+sql: "SELECT my_volatile()"
+options:
+  escalate_unknown_functions: true
+  safe_function_allowlist: [now, current_timestamp]
+expected_classification:
+  - raw_verb: SELECT
+    primary_group: procedural
+    effects:
+      - group: procedural
+      - group: read
